Projects/fivetran_connectors/fivetran/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pymssql==2.1.4
PyYAML==5.3
requests==2.22.0
urllib3==1.25.8